@@ -1,6 +1,6 @@
 package com.udacity.notepad.data;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.Date;
 
