@@ -2,8 +2,8 @@ package com.udacity.notepad.util;
 
 import android.content.Context;
 import android.graphics.Rect;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public class SpaceItemDecoration extends RecyclerView.ItemDecoration {
